requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
onnxruntime>=1.16.0
pydantic>=2.5.0
passlib[bcrypt]
python-dotenv~=1.2.1
bcrypt==4.3.0
python-jose==3.3.0
python-multipart>=0.0.5
